Dz3 fix mistakes
